- fix AddField typo - use AfterApplyDefaultBehaviour to move timing of type configuration (moves config code to GrpcConfig) - above needs newer protobuf-net build
