+++ src/ServiceStack.Extensions/ServiceStack.Extensions.csproj
@@ -12,7 +12,7 @@
     <ItemGroup>
         <PackageReference Include="Grpc.AspNetCore.Server" Version="2.28.0-pre2" />
         <PackageReference Include="Grpc.Core.Api" Version="2.28.0-pre3" />
-        <PackageReference Include="protobuf-net" Version="3.0.0-alpha.133" />
+        <PackageReference Include="protobuf-net" Version="3.0.0-alpha.143" />
         <PackageReference Include="protobuf-net.Grpc.AspNetCore" Version="1.0.36" />
         
         <PackageReference Include="System.Threading.Channels" Version="4.7.0" />
