+++ src/ServiceStack.GrpcClient/ServiceStack.GrpcClient.csproj
@@ -13,7 +13,7 @@
         <PackageReference Include="Grpc.Core" Version="2.26.0" />
         <PackageReference Include="protobuf-net.Grpc" Version="1.0.22" />
         <!-- the transitive import of protobuf-net is lower; we need at least this -->
-        <PackageReference Include="protobuf-net" Version="2.4.5" />
+        <PackageReference Include="protobuf-net" Version="2.4.6" />
         <PackageReference Include="ServiceStack.Text" Version="$(Version)" />
 
         <ProjectReference Include="..\ServiceStack.Interfaces\ServiceStack.Interfaces.csproj" />
