+++ src/ServiceStack.ProtoBuf/ServiceStack.ProtoBuf.Core.csproj
@@ -16,6 +16,6 @@
     <ProjectReference Include="..\ServiceStack.Common\ServiceStack.Common.Core.csproj" />
     <ProjectReference Include="..\ServiceStack\ServiceStack.Core.csproj" />
     <PackageReference Include="ServiceStack.Text.Core" Version="$(Version)" />
-    <PackageReference Include="protobuf-net" Version="2.4.5" />
+    <PackageReference Include="protobuf-net" Version="2.4.6" />
   </ItemGroup>
 </Project>
